Inicio das Telas - Swing
Implementado a Tela do Paciente

+++ java/untitled/src/main/java/model/Paciente.java
@@ -10,7 +10,7 @@ public class Paciente extends Pessoa {
     private String nomeCuidador;
     private String telefoneCuidador;
 
-    public Paciente(int id, String nome, String cpf, String celular, Endereco endereco, String dataNascimento, String cns, String email, String nomeCuidador, String telefoneCuidador) {
+    public Paciente(Long id, String nome, String cpf, String celular, Endereco endereco, String dataNascimento, String cns, String email, String nomeCuidador, String telefoneCuidador) {
         super(id, nome, cpf, celular);
         this.endereco = endereco;
         this.dataNascimento = LocalDate.parse(dataNascimento);
